moved offering extensions to standard activation service
