+++ core/api/src/main/java/org/n52/sos/config/SosActivationDao.java
@@ -30,9 +30,8 @@ package org.n52.sos.config;
 
 import java.util.Set;
 
-import org.n52.sos.coding.encode.ProcedureDescriptionFormatKey;
 import org.n52.iceland.coding.encode.ResponseFormatKey;
-import org.n52.iceland.ogc.swes.OfferingExtensionKey;
+import org.n52.sos.coding.encode.ProcedureDescriptionFormatKey;
 
 /**
  *
@@ -93,17 +92,4 @@ public interface SosActivationDao {
 
     Set<ProcedureDescriptionFormatKey> getProcedureDescriptionFormatKeys();
 
-    /**
-     * Checks if the offering extension is active.
-     *
-     * @param key
-     *            the offering extension key
-     *
-     * @return if the offering extension is active
-     */
-    boolean isOfferingExtensionActive(OfferingExtensionKey key);
-
-    void setOfferingExtensionStatus(OfferingExtensionKey key, boolean active);
-
-    Set<OfferingExtensionKey> getOfferingExtensionKeys();
 }
